+++ EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -4,7 +4,7 @@
 
 import CoreData
 
-public final class CoreDataFeedStore: FeedStore {
+public final class CoreDataFeedStore {
 	public static let modelName = "FeedStore"
 	public static let model = NSManagedObjectModel(name: modelName, in: Bundle(for: CoreDataFeedStore.self))
 
@@ -36,47 +36,6 @@ public final class CoreDataFeedStore: FeedStore {
 		context.performAndWait {
 			let coordinator = self.container.persistentStoreCoordinator
 			try? coordinator.persistentStores.forEach(coordinator.remove)
-		}
-	}
-
-	public func retrieve(completion: @escaping RetrievalCompletion) {
-		perform { context in
-            completion(Result {
-                try ManagedCache.find(in: context).map {
-                    CachedFeed(feed: $0.localFeed, timestamp: $0.timestamp)
-                }
-            })
-		}
-	}
-
-	public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		perform { context in
-			do {
-				let managedCache = try ManagedCache.newUniqueInstance(in: context)
-				managedCache.timestamp = timestamp
-				managedCache.feed = ManagedFeedImage.images(from: feed, in: context)
-
-				try context.save()
-                completion(.success(()))
-			} catch {
-				context.rollback()
-                completion(.failure(error))
-			}
-		}
-	}
-
-	public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		perform { context in
-			do {
-				try ManagedCache.find(in: context)
-					.map(context.delete)
-					.map(context.save)
-
-                completion(.success(()))
-			} catch {
-				context.rollback()
-                completion(.failure(error))
-			}
 		}
 	}
